requirements.txt
google-generativeai>=0.7.2
feedparser
python-telegram-bot
python-dotenv